Update GitHub CI workflows and add draft check
This update includes changes in the workflow names and adds a new job to check the draft state of a pull request. This job ensures to run the workflows only if the pull request is not in draft state. Also, it introduces dependencies with this new draft job in both dockers and host jobs.